refactor: class명 변경 & 메소드 기능별 분리
